chore: :truck: create blog-post folder under templates folder
